ai-assistant/requirements.txt
# WebRTC and networking
aiortc>=1.14.0
aiohttp>=3.13.2
aiohttp_cors>=0.8.1
websockets>=15.0.1  # Used in test client only

# Google Cloud APIs with gRPC support
google-cloud-speech>=2.34.0
google-cloud-texttospeech>=2.33.0
grpcio>=1.60.0
grpcio-status>=1.60.0

# LangChain and LLM integration
langchain>=0.3.0
langchain-google-genai>=2.0.0
langchain-community>=0.3.0
langchain-core>=0.3.0

# Weaviate Vector Database
weaviate-client>=4.9.0

# Audio processing
numpy>=2.3.4

# Environment variables
python-dotenv>=1.2.1

# Firebase Authentication
firebase-admin>=6.5.0

# Testing dependencies
pytest>=8.0.0
pytest-aiohttp>=1.0.5
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0
